fix: match the one-hot target's dtype to the detection logits
Under autocast the FCOS class logits are bf16 or fp16, so the one-hot
tensor allocated beside them is too. Forcing the source to float32 made
index_put_ reject it:

    RuntimeError: Index put requires the source and destination dtypes
    match, got BFloat16 for the destination and Float for the source.

Reaching the line needs three things at once: AMP enabled, a CUDA-class
device, and the detection head actually supervised. Nothing supplied the
third until COCO was added to the indoor config, so a full 60-epoch run
and 166 green tests never executed it once.

The regression test forces autocast explicitly instead of reading the
device, because supports_amp() disables AMP on CPU and MPS and a
device-dependent test would run vacuous in CI -- which is part of how
this survived. Reverting the cast fails 8 of its 9 cases; the one that
still passes is the fp32 path, which is the boundary being asserted.

+++ src/syncai_hydranet/models/losses.py
@@ -108,7 +108,11 @@ class FCOSLoss(nn.Module):
         pos = cls_t < self.num_classes
         num_pos = pos.sum().clamp(min=1).float()
         onehot = torch.zeros_like(flat_cls)
-        onehot[pos] = F.one_hot(cls_t[pos], self.num_classes).float()
+        # Match flat_cls's dtype rather than forcing float32: under autocast the
+        # logits are bf16/fp16, and index_put_ refuses a mismatched source. This
+        # only fires when the detection head is actually supervised on CUDA, which
+        # is why it survived every seg-only run.
+        onehot[pos] = F.one_hot(cls_t[pos], self.num_classes).to(onehot.dtype)
         cls_loss = sigmoid_focal_loss(flat_cls, onehot) / num_pos
         if pos.any():
             reg_loss = giou_loss(flat_reg[pos], reg_t[pos]) / num_pos
